Web Shell
This web shell is designed for gaining root on a server. The shell has a login prompt, view contents of server, server over-view, user over-view, execute server commands on the server and write/upload PHP code to the server. The shell is made to give the user root privileges and execute commands that only the root can execute.  

Login: Twemlow
Password: Emrys

The source code is obfuscated however main source is on my GitHub. The shell was developed on a local Apache server using PHP 8.0.2.